Made Hangman Draw its Image
Finally! Turns out I had to wait until the
window loaded so that I could draw it.

Anyway, the problem now is to get the image
not to be so blurry, I guess in the future
I could use something like an svg instead
of an image element to solve it.