Made Hangman Work
I fixed the blurryness problem by setting
the width and height of the canvas to be the
same as those of the image (manually, but I
can make it automatically; I will in the
future). The hangman now can advance stage
properly.

Also, fixed a few problems in the color class
(especifically toHex(number)), and added
the method sub(Color, Color)