Set Up Game Methods
---
Set up basics for the game, focusing on the
ability to restart without reloading the page.

@scripts/hangman.js	Separated the functions
of the hangman so that I can reset its
colors on another game/round without
repainting the hangman image

@scripts/letter-space-panel.js	Class that will
hold the spaces for the letters of a word to
guess

@scripts/word.js	Fixed many things that
have to do with getting a random word from
the API; turns out I don't know as much as
I thought about promises and async/await.

@main.html			Added a new feature;
a HUD to indicate data about the player's game

@main.js			Added the whole bunch
of functions (most of them not finished yet)
that I will use to make the game work properly.

@TODO
* Set styling for the letter-space-panel
(currently, the panel doesn't show at first
and the letters pile up vertically)

* Finish the functions to keep track of the game
(also, decide if you are going to execute
`onNewGame` right away, or still repeat the
code at the beginning)